requirements.txt
livekit==0.22.0
livekit-agents==0.12.17
livekit-api==0.8.2
livekit-plugins-deepgram==0.7.1
livekit-plugins-elevenlabs==0.8.1
livekit-plugins-google==0.11.1
livekit-plugins-openai==0.11.3
livekit-plugins-silero==0.7.5
livekit-protocol==0.9.2
mem0ai==0.1.91
python-dotenv==1.0.1